Very much non-compiling checkpoint
